fix(new-wallet): rename doublecheck content folder
