+++ src/modules/NewWallet/DoubleCheck/DoubleCheckContent/WordsCheckerController.tsx
