+++ src/modules/NewWallet/DoubleCheck/DoubleCheckContent/index.tsx
